Added files via upload
Current progress on front-end. Still need to add UMBC CSEE address and Contact link to div.footer then get div.container to scale properly to text in div.article, but otherwise the base is almost ready to go. The bright red and blue backgrounds are just for temporary testing to see div sections and are intended to be white or off-white once scaling issues are resolved.

"Home" page is intended to be a brief one or two paragraphs describing what Azimuth does to the visitor, and if you guys want to stick with the name, a small sidebar that explains what an azimuth is and how it relates to the website ("navigating" through degree course requirements). "Get Started" page is where the first user input is handled, from where we will link over to second and third input / output pages (course selection, course output). "About" is just a short one or two paragraph page describing the project with an additional link to Contact info.